refactor: split code into ES modules
