+++ index.html
@@ -105,6 +105,6 @@
     </div>
   </div>
 
-  <script src="js/script.js" defer></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
