+++ js/ui/ui.js
@@ -1,292 +1,8 @@
-// ----------------------------
-// Helper: Returns an array of holiday dates for a given year
-// ----------------------------
-function getHolidays(year) {
-    return [
-        new Date(year, 0, 1),    // New Year's Day
-        new Date(year, 4, 1),    // Labour Day
-        new Date(year, 11, 25),  // Christmas Day
-        new Date(year, 0, 6),    // Reyes
-        new Date(year, 2, 3),    // Carnival
-        new Date(year, 2, 4),    // Carnival
-        new Date(year, 3, 18),   // 33 - moved
-        new Date(year, 5, 19),   // Artigas
-        new Date(year, 6, 18),   // Constitution Day
-        new Date(year, 7, 25)    // Independence Day
-    ];
-}
+import { OptionsManager } from './optionsManager.js';
+import { getHolidays } from '../attendance/utils.js';
+import { showAttendanceNotification, calculateDistance, themes } from '../helpers.js';
 
-// ----------------------------
-// Helper function to calculate working days in a month
-// (Monday-Friday excluding fixed holidays)
-// ----------------------------
-function calculateWorkingDays(year, month) {
-    const holidays = getHolidays(year);
-    const date = new Date(year, month, 1);
-    let count = 0;
-    while (date.getMonth() === month) {
-        const dayOfWeek = date.getDay();
-        const isHoliday = holidays.some(holiday =>
-            holiday.getDate() === date.getDate() &&
-            holiday.getMonth() === date.getMonth() &&
-            holiday.getFullYear() === date.getFullYear()
-        );
-        if (dayOfWeek !== 0 && dayOfWeek !== 6 && !isHoliday) count++;
-        date.setDate(date.getDate() + 1);
-    }
-    return count;
-}
-
-// ----------------------------
-// Debounce Helper Function
-// ----------------------------
-function debounce(func, wait) {
-    let timeout;
-    return function(...args) {
-        const context = this;
-        clearTimeout(timeout);
-        timeout = setTimeout(() => func.apply(context, args), wait);
-    };
-}
-
-// ----------------------------
-// Notification Helper Function
-// ----------------------------
-/**
- * Shows a browser notification after requesting permission.
- * @param {string} body The text to display in the notification.
- * @param {string} iconUrl Optional URL for an icon.
- */
-function showAttendanceNotification(body, iconUrl = 'data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><rect width=%2290%22 height=%2290%22 x=%225%22 y=%225%22 fill=%22%23fff%22 stroke=%22%23000%22 stroke-width=%2210%22/><path d=%22M30 50 L45 65 L70 35%22 stroke=%22%234caf50%22 stroke-width=%2210%22 fill=%22none%22/></svg>') {
-    if (!("Notification" in window)) {
-        console.error("This browser does not support desktop notification.");
-        return;
-    }
-
-    if (Notification.permission === "granted") {
-        new Notification("Attendance Tracker", { body: body, icon: iconUrl });
-    } else if (Notification.permission !== "denied") {
-        Notification.requestPermission().then(permission => {
-            if (permission === "granted") {
-                new Notification("Attendance Tracker", { body: body, icon: iconUrl });
-            }
-        });
-    }
-}
-
-// ----------------------------
-// Geolocation Distance Helper Function
-// ----------------------------
-/**
- * Calculates the distance between two lat/lon coordinates in meters.
- * @param {{latitude: number, longitude: number}} loc1 First location.
- * @param {{latitude: number, longitude: number}} loc2 Second location.
- * @returns {number} Distance in meters.
- */
-function calculateDistance(loc1, loc2) {
-    const R = 6371e3; // metres
-    const φ1 = loc1.latitude * Math.PI / 180;
-    const φ2 = loc2.latitude * Math.PI / 180;
-    const Δφ = (loc2.latitude - loc1.latitude) * Math.PI / 180;
-    const Δλ = (loc2.longitude - loc1.longitude) * Math.PI / 180;
-    const a = Math.sin(Δφ / 2) * Math.sin(Δφ / 2) +
-              Math.cos(φ1) * Math.cos(φ2) *
-              Math.sin(Δλ / 2) * Math.sin(Δλ / 2);
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    return R * c; // in metres
-}
-
-// ----------------------------
-// Global Themes Variable
-// ----------------------------
-const themes = {
-    'default': { background: '#fffbf7', foreground: '#45372b', accent: '#df7020' },
-    'bsod': { background: '#153489', foreground: '#eceae5', accent: '#5ea5ee' },
-    'dracula': { background: '#282a36', foreground: '#f8f8f2', accent: '#f44336' },
-    'ultra-violet': { background: '#440184', foreground: '#BF00FF', accent: '#E78FFF' },
-    'hello-kitty': { background: '#FFF0F5', foreground: '#4a4a4a', accent: '#ff1493' },
-    'matrix': { background: '#2b2b2b', foreground: '#4eee85', accent: '#4eee85' },
-    'dark-mode': { background: '#303030', foreground: '#e0e0e0', accent: '#9e9e9e' },
-    'windows-98': { background: '#c0c0c0', foreground: '#000000', accent: '#008080' },
-    'mint': { background: '#e5ffe5', foreground: '#2e8b57', accent: '#32cd3f' },
-    'terminal': { background: '#1a170f', foreground: '#eceae5', accent: '#eec35e' }
-};
-
-// ----------------------------
-// AttendanceHistory Class: Handles multi-month data in localStorage
-// ----------------------------
-class AttendanceHistory {
-    constructor() {
-        try {
-            const data = localStorage.getItem('attendanceHistory');
-            this.data = data ? JSON.parse(data) : {};
-        } catch (error) {
-            console.error("Error reading attendance history from localStorage", error);
-            this.data = {};
-        }
-    }
-    getMonthKey(year, month) {
-        const m = month < 10 ? '0' + month : month;
-        return `${year}-${m}`;
-    }
-    getMonthData(year, month) {
-        const key = this.getMonthKey(year, month);
-        return this.data[key] || [];
-    }
-    saveMonthData(year, month, entries) {
-        const key = this.getMonthKey(year, month);
-        this.data[key] = entries;
-        try {
-            localStorage.setItem('attendanceHistory', JSON.stringify(this.data));
-        } catch (error) {
-            console.error("Error saving attendance history to localStorage", error);
-        }
-    }
-    clearMonthData(year, month) {
-        const key = this.getMonthKey(year, month);
-        delete this.data[key];
-        try {
-            localStorage.setItem('attendanceHistory', JSON.stringify(this.data));
-        } catch (error) {
-            console.error("Error clearing attendance history in localStorage", error);
-        }
-    }
-}
-
-// ----------------------------
-// AttendanceManager Class: Handles data for the current month
-// ----------------------------
-class AttendanceManager {
-    constructor(year, month, history) {
-        this.year = year;
-        this.month = month; // 0-indexed (for date calculations)
-        this.monthForStorage = month + 1;
-        this.attendanceHistory = history;
-        this.attendanceDays = this.loadAttendance();
-        this.workingDays = calculateWorkingDays(year, month);
-        // Default attendance goal is set to 55% (can be updated via the slider)
-        this.requiredAttendance = Math.ceil(this.workingDays * 0.55);
-    }
-    loadAttendance() {
-        return this.attendanceHistory.getMonthData(this.year, this.monthForStorage);
-    }
-    saveAttendance() {
-        this.attendanceHistory.saveMonthData(this.year, this.monthForStorage, this.attendanceDays);
-    }
-    hasAttendance(day) {
-        return this.attendanceDays.some(entry => entry.day === day);
-    }
-    addAttendance(day) {
-        if (!this.hasAttendance(day)) {
-            this.attendanceDays.push({ day, month: this.monthForStorage, year: this.year });
-            this.saveAttendance();
-        }
-    }
-    removeAttendance(day) {
-        const index = this.attendanceDays.findIndex(entry => entry.day === day);
-        if (index !== -1) {
-            this.attendanceDays.splice(index, 1);
-            this.saveAttendance();
-        }
-    }
-    toggleAttendance(day) {
-        if (this.hasAttendance(day)) {
-            this.removeAttendance(day);
-        } else {
-            this.addAttendance(day);
-        }
-    }
-    clearAttendance() {
-        this.attendanceDays = [];
-        this.attendanceHistory.clearMonthData(this.year, this.monthForStorage);
-    }
-    getCount() {
-        return this.attendanceDays.length;
-    }
-}
-
-// ----------------------------
-// OptionsManager Class
-// ----------------------------
-class OptionsManager {
-    constructor() {
-        this.storageKey = 'customOptions';
-        this.options = {
-            background: getComputedStyle(document.documentElement).getPropertyValue('--color-bg').trim(),
-            foreground: getComputedStyle(document.documentElement).getPropertyValue('--color-text').trim(),
-            accent: getComputedStyle(document.documentElement).getPropertyValue('--color-progress-fill').trim(),
-            attendanceGoal: 55
-        };
-        this.loadOptions();
-    }
-    loadOptions() {
-        const stored = localStorage.getItem(this.storageKey);
-        if (stored) {
-            try {
-                const parsed = JSON.parse(stored);
-                this.options = { ...this.options, ...parsed };
-                this.applyOptions();
-            } catch (e) {
-                console.error("Error parsing options from localStorage", e);
-            }
-        }
-    }
-    saveOptions() {
-        try {
-            localStorage.setItem(this.storageKey, JSON.stringify(this.options));
-        } catch (error) {
-            console.error("Error saving options to localStorage", error);
-        }
-    }
-    applyOptions() {
-        document.documentElement.style.setProperty('--color-bg', this.options.background);
-        document.documentElement.style.setProperty('--color-card-bg', this.options.background);
-        document.documentElement.style.setProperty('--color-calendar-bg', this.options.background);
-        document.documentElement.style.setProperty('--color-btn-bw-bg', this.options.background);
-        document.documentElement.style.setProperty('--color-overlay-bg', this.options.background);
-
-        document.documentElement.style.setProperty('--color-text', this.options.foreground);
-        document.documentElement.style.setProperty('--color-border', this.options.foreground);
-        document.documentElement.style.setProperty('--color-overlay-text', this.options.foreground);
-
-        document.documentElement.style.setProperty('--color-progress-fill', this.options.accent);
-        document.documentElement.style.setProperty('--color-btn-bw-hover', this.options.accent);
-
-        function hexToRgb(hex) {
-            hex = hex.replace(/^#/, '');
-            if (hex.length === 3) {
-                hex = hex.split('').map(c => c + c).join('');
-            }
-            const bigint = parseInt(hex, 16);
-            const r = (bigint >> 16) & 255;
-            const g = (bigint >> 8) & 255;
-            const b = bigint & 255;
-            return { r, g, b };
-        }
-        function rgbToHex(r, g, b) {
-            return '#' + [r, g, b].map(x => {
-                const hex = x.toString(16);
-                return hex.length === 1 ? '0' + hex : hex;
-            }).join('');
-        }
-        const baseRgb = hexToRgb(this.options.background);
-        const newR = Math.max(0, Math.min(255, baseRgb.r - 19));
-        const newG = Math.max(0, Math.min(255, baseRgb.g - 15));
-        const newB = Math.max(0, Math.min(255, baseRgb.b - 11));
-        const newProgressBg = rgbToHex(newR, newG, newB);
-        document.documentElement.style.setProperty('--color-progress-bg', newProgressBg);
-    }
-    updateOption(key, value) {
-        this.options[key] = value;
-        this.saveOptions();
-        this.applyOptions();
-    }
-}
-
-// ----------------------------
-// AttendanceUI Class: Manages ALL UI updates and event bindings
-// ----------------------------
-class AttendanceUI {
+export class AttendanceUI {
     constructor(manager, today) {
         this.manager = manager;
         this.today = today;
@@ -838,90 +554,3 @@ class AttendanceUI {
         }
     }
 }
-
-// ----------------------------
-// Initialization Code
-// ----------------------------
-const today = new Date();
-const year = today.getFullYear();
-const month = today.getMonth(); // 0-indexed
-const attendanceHistory = new AttendanceHistory();
-const attendanceManager = new AttendanceManager(year, month, attendanceHistory);
-const attendanceUI = new AttendanceUI(attendanceManager, today);
-
-// ----------------------------
-// Auto-reload if data is older than 60 minutes
-// ----------------------------
-const reloadIntervalMinutes = 60;
-localStorage.setItem("lastReload", Date.now());
-document.addEventListener("visibilitychange", () => {
-    if (document.visibilityState === "visible") {
-        const lastReload = parseInt(localStorage.getItem("lastReload"), 10);
-        const now = Date.now();
-        const minutesPassed = (now - lastReload) / (1000 * 60);
-        if (minutesPassed > reloadIntervalMinutes) {
-            localStorage.setItem("lastReload", now);
-            location.reload();
-        } else {
-            console.log("Data is still fresh, no reload needed");
-        }
-    }
-});
-
-// ----------------------------
-// Optional Helper Functions for Testing
-// ----------------------------
-window.clearAttendance = () => {
-    attendanceManager.clearAttendance();
-    attendanceUI.updateAll();
-};
-window.exportAttendance = () => {
-    attendanceHistory.data && Object.values(attendanceHistory.data).forEach(monthEntries => {
-        monthEntries.forEach(entry => {
-            console.log(`Day: ${entry.day}, Month: ${entry.month}, Year: ${entry.year}`);
-        });
-    });
-};
-window.addAttendanceMark = (day, m, y) => {
-    const date = new Date(y, m, day);
-    if (date.getMonth() === m && date.getDate() === day && date.getFullYear() === y) {
-        const monthForStorage = m + 1;
-        const entries = attendanceHistory.getMonthData(y, monthForStorage);
-        if (!entries.some(d => d.day === day)) {
-            entries.push({ day, month: monthForStorage, year: y });
-            attendanceHistory.saveMonthData(y, monthForStorage, entries);
-            attendanceUI.updateAll();
-        } else {
-            console.log("This date is already marked.");
-        }
-    } else {
-        console.log("Invalid date.");
-    }
-};
-window.addAttendanceDays = function (days) {
-    for (let i = 0; i < days; i++) {
-        if (attendanceManager.getCount() < attendanceManager.workingDays) {
-            const date = new Date(attendanceManager.year, attendanceManager.month, attendanceManager.getCount() + 1);
-            attendanceManager.addAttendance(date.getDate());
-        }
-    }
-    attendanceUI.updateAll();
-};
-window.clearAllAttendance = function () {
-    localStorage.removeItem('attendanceHistory');
-    console.log('All attendance data cleared! Fresh start achieved.');
-    location.reload();
-};
-window.createFakeMarkData = function () {
-    const year = new Date().getFullYear();
-    for (let m = 1; m <= 12; m++) {
-        const daysInMonth = new Date(year, m, 0).getDate();
-        const randomDays = new Set();
-        while (randomDays.size < 10) {
-            randomDays.add(Math.floor(Math.random() * daysInMonth) + 1);
-        }
-        randomDays.forEach(day => {
-            window.addAttendanceMark(day, m - 1, year);
-        });
-    }
-};
